+++ tests/bootstrap.php
@@ -8,11 +8,13 @@ session_start();
 $logger = Configuration::getLogger();
 $logger->info("Installing database from ". Configuration::getConfigFile());
 
+$doInstall = getenv("INSTALL");
+
 // drop all content if exists
 $cdb = new CoreDB();
-$cdb->dropAll();
+$cdb->dropAll(drop: $doInstall !== "0");
 
-$doInstall = getenv("INSTALL");
+
 if($doInstall !== false && $doInstall === "0") {
     $m = new CoreUser();
     $m->installDefault();